backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.0
django-cors-headers==4.3.1
requests==2.31.0
python-decouple==3.8
azure-ai-inference==1.0.0b9
azure-core==1.30.0 